Extract integration test cases to a module
In order to reuse the same test logic for different event backends, we want to
extract the generic test logic to a module that can be reused.
